+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessServiceTest.scala
@@ -93,6 +93,7 @@ class PeriodicProcessServiceTest
       },
       DeploymentRetryConfig(),
       PeriodicExecutionConfig(),
+      maxFetchedPeriodicScenarioActivities = None,
       new ProcessConfigEnricher {
 
         override def onInitialSchedule(